Add method NamespacedKV.Subspace
This will be useful for storing WebAuthn credentials as individual records
instead of in one blob. This way the `webauthnService` can receive a subspace of
the KV store and operate on that with just credential IDs as keys, instead of
having to remember to add the subspace prefix in every individual call.